refactor: GameSystem에서 입력 및 입력값 검증 로직 분리
